# backend-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "School Management System API",
        "description": "API for school management system including user authentication, and resource management",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/tbdrs-merge": {
            "get": {
                "tags": [
                    "TBDRS Merge"
                ],
                "summary": "Get all TBDRS merges",
                "operationId": "a6c8ed7c75af6b3bb9907a69c1d3c80c",
                "responses": {
                    "200": {
                        "description": "List of TBDRS merges with relationships",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tbdrs_merges": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "track_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "building_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "department_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "section_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "track": {
                                                        "type": "object"
                                                    },
                                                    "building": {
                                                        "type": "object"
                                                    },
                                                    "department": {
                                                        "type": "object"
                                                    },
                                                    "section": {
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "TBDRS merges retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "TBDRS Merge"
                ],
                "summary": "Create a new TBDRS merge",
                "operationId": "9bbd32aa5cb2f66e369c18273cff83bd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "track_id",
                                    "building_id",
                                    "department_id",
                                    "section_id"
                                ],
                                "properties": {
                                    "track_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "building_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "department_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "section_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "TBDRS merge created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tbdrs_merge": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "TBDRS merge created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "404": {
                        "description": "One or more referenced entities not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tbdrs-merge/{id}": {
            "get": {
                "tags": [
                    "TBDRS Merge"
                ],
                "summary": "Get a specific TBDRS merge",
                "operationId": "ee25742a55b2474f2e68bc51e39984cb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "TBDRS Merge ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "TBDRS merge details with relationships",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tbdrs_merge": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "TBDRS merge retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "TBDRS merge not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "TBDRS Merge"
                ],
                "summary": "Update a TBDRS merge",
                "operationId": "afdc0ad532c9303c1eb1db403b42c00b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "TBDRS Merge ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "track_id",
                                    "building_id",
                                    "department_id",
                                    "section_id"
                                ],
                                "properties": {
                                    "track_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "building_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "department_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "section_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "TBDRS merge updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tbdrs_merge": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "TBDRS merge updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "TBDRS merge not found"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "TBDRS Merge"
                ],
                "summary": "Delete a TBDRS merge",
                "operationId": "3cfbcf31fa533878929a78ce8141bf25",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "TBDRS Merge ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "TBDRS merge deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "TBDRS merge deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "TBDRS merge not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}